% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_tag.R
\name{find_tag}
\alias{find_tag}
\alias{find_tag_anime}
\alias{find_tag_manga}
\alias{find_tag_people}
\alias{find_tag_character}
\title{Find the correct name for the search item}
\usage{
find_tag(name = NULL, sourceType = "")

find_tag_anime(name)

find_tag_manga(name)

find_tag_people(name)

find_tag_character(name)
}
\arguments{
\item{name}{the MAL name of interest}

\item{sourceType}{the type of medium of the search i.e. manga, anime, character, person}
}
\value{
returns the index number for the search item; combines with general_info family of functions to gather the actual data
}
\description{
Find the correct name for the search item
}
\examples{

\dontrun{
find_tag('bakemonogatari')
find_tag_anime('naruto')
find_tag_manga('Opus')
find_tag_character('Tomie')
find_tag_people('hana kana')
}

}
