% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_tag.R
\name{find_tag}
\alias{find_tag}
\title{Find the correct name for the search item}
\usage{
find_tag(name = NULL, sourceType = "")
}
\arguments{
\item{name}{the MAL name of interest}

\item{sourceType}{the type of medium of the search i.e. manga, anime, character, person}
}
\value{
returns the index number for the search item; combines with info_general functions to gather the actual data
}
\description{
Find the correct name for the search item
}
