% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_info.R
\name{general_info}
\alias{general_info}
\alias{anime_info}
\alias{manga_info}
\alias{character_info}
\alias{people_info}
\title{Find MAL data}
\usage{
general_info(name = NULL, type = NULL, auto = FALSE)

anime_info(name, type = "anime", ...)

manga_info(name, type = "manga", ...)

character_info(name, type = "character", ...)

people_info(name, type = "people", ...)
}
\arguments{
\item{name}{the MAL name of interest}

\item{type}{the type of medium of the search i.e. manga, anime, character, person}
}
\value{
the results from the search in a dataframe format
}
\description{
Find MAL data
}
\examples{
\dontrun{
general_info('naruto', type = 'anime')
anime_info('bakemonogatari')
manga_info('Opus')
character_info('Tomie')
people_info('hana kana')
}
}
