% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anime_info.R
\name{info_general}
\alias{info_general}
\title{Find MAL data}
\usage{
info_general(name = NULL, type = NULL)
}
\arguments{
\item{name}{the MAL name of interest}

\item{type}{the type of medium of the search i.e. manga, anime, character, person}
}
\value{
the results from the search
}
\description{
Find MAL data
}
