% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jikanr-package.R
\docType{package}
\name{jikanr-package}
\alias{jikanr}
\alias{jikanr-package}
\title{jikanr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Kyle Harris \email{koderkow@gmail.com}

}
\keyword{internal}
