% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{schedule}
\alias{schedule}
\title{Anime of the specified day}
\usage{
schedule(day = NULL)
}
\arguments{
\item{day}{Monday through Friday. Defaults to NULL and returns a little of all days.}
}
\description{
Anime schedule of the week or specified day. Note: If you don't pass the day parameter, it'll return the schedule for all days of the week.
}
\examples{
schedule("monday")

}
