% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.R
\name{season}
\alias{season}
\title{Anime of the specified season and year}
\usage{
season(year, season)
}
\arguments{
\item{year}{Specify the year.}

\item{season}{Specify the season: summer, spring, fall, winter.}
}
\description{
Anime of the specified season. Note: Both parameters are required.
}
\examples{
season(1999, "winter")

}
