% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top.R
\name{top}
\alias{top}
\title{Top items on MyAnimeList. Note: subtype returns a filtered top list of a type
item. For example, the top Anime (type) movies (subtype). Note2: subtype is
only for anime and manga types.}
\usage{
top(type, subtype = NULL, page = 1)
}
\arguments{
\item{type}{Top items of this type}

\item{subtype}{Refine the data using a subtype such as: \cr
Anime: airing, upcoming, tv, movie, ova, special. \cr
Manga: manga, novels, oneshots, doujin, manhwa, manhua, \cr
Both: bypopularity, favorite.}

\item{page}{The Top page on MyAnimeList is paginated offers 50 items per page}
}
\description{
Top items on MyAnimeList. Note: subtype returns a filtered top list of a type
item. For example, the top Anime (type) movies (subtype). Note2: subtype is
only for anime and manga types.
}
