% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jikanr-package.R
\docType{package}
\name{jikanr-package}
\alias{jikanr}
\alias{jikanr-package}
\title{jikanr: R wrapper for the Jikan API: The Unofficial MyAnimeList API}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KoderKow/jikanr}
  \item Report bugs at \url{https://github.com/KoderKow/jikanr/issues}
}

}
\author{
\strong{Maintainer}: Kyle Harris \email{koderkow@gmail.com}

}
\keyword{internal}
