% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Query items on MyAnimeList.}
\usage{
query(type, query, page = 1)
}
\arguments{
\item{type}{Top items of this type.}

\item{query}{A string containing the query, example: "dragon ball".}

\item{page}{The Top page on MyAnimeList is paginated offers 50 items per page.}
}
\description{
Search results for the query. NOTE: MyAnimeList only processes queries with a minimum of 3 letters.
}
\examples{
query("manga", "berserk")

}
